near-real-time-demo-docker/requirements.txt
datasets
huggingface_hub
lightning
matplotlib
numpy
opencv-python==4.5.5.64
Pillow
pytorch_lightning
streamlit
torch
torchmetrics
torchvision
transformers
